+++ src/mnelab/dialogs/history.py
@@ -13,7 +13,7 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
 )
 
-from mnelab.utils import PythonHighlighter
+from mnelab.utils import PythonHighlighter, format_code
 
 
 class HistoryDialog(QDialog):
@@ -34,14 +34,15 @@ class HistoryDialog(QDialog):
         text.setFont(font)
         highlighter = PythonHighlighter(text.document())  # noqa: F841
         text.setReadOnly(True)
+        history = format_code(history)
         text.setPlainText(history)
         layout.addWidget(text)
         buttonbox = QDialogButtonBox(QDialogButtonBox.Ok)
         clipboardbutton = QPushButton("Copy to clipboard")
         buttonbox.addButton(clipboardbutton, QDialogButtonBox.ActionRole)
         clipboard = QGuiApplication.clipboard()
-        clipboardbutton.clicked.connect(lambda: clipboard.setText(history + "\n"))
+        clipboardbutton.clicked.connect(lambda: clipboard.setText(history))
         layout.addWidget(buttonbox)
         self.setLayout(layout)
         buttonbox.accepted.connect(self.accept)
-        self.resize(700, 500)
+        self.resize(750, 500)